fix(platform): remove circular dependencies
