+++ libs/platform/src/lib/form/form-group/form-group-header/form-group-header.component.ts
@@ -1,7 +1,7 @@
 import { ChangeDetectionStrategy, Component, Input, ViewEncapsulation } from '@angular/core';
-import { FieldGroup } from '../../form-helpers';
 import { HintOptions } from '@fundamental-ngx/platform/shared';
 import { Nullable } from '@fundamental-ngx/cdk/utils';
+import { FieldGroup } from '../../models/field.model';
 
 @Component({
     // eslint-disable-next-line @angular-eslint/component-selector
