+++ libs/platform/src/lib/form/form-helpers.ts
@@ -1,13 +1,8 @@
-import { TemplateRef } from '@angular/core';
-
-import { PlatformFormField, HintOptions } from '@fundamental-ngx/platform/shared';
+import { PlatformFormField } from '@fundamental-ngx/platform/shared';
 import { FormFieldComponent } from './form-group/form-field/form-field.component';
 import { FormFieldGroupComponent } from './form-group/form-field-group/form-field-group.component';
 import { FormGeneratorFieldComponent } from './form-generator/form-generator-field/form-generator-field.component';
-
-export interface FieldColumn {
-    [key: number]: Array<Field>;
-}
+import { Field } from './models/field.model';
 
 /** @hidden */
 export function isFieldChild(child: unknown): child is PlatformFormField {
@@ -34,19 +29,4 @@ export function getField(field: PlatformFormField): Field {
     field = isFieldGroupWrapperChild(field) ? field.fieldRenderer : field;
 
     return new Field(field.id, field.rank, field.renderer, field.column);
-}
-
-export class Field {
-    /** @hidden */
-    constructor(
-        public name?: string,
-        public rank?: number,
-        public renderer?: TemplateRef<any>,
-        public column?: number
-    ) {}
-}
-
-export class FieldGroup {
-    /** @hidden */
-    constructor(public label: string, public fields: FieldColumn, public hintOptions?: HintOptions) {}
 }
